Review fixes: num->int/double conversions on clamp() results, LRU byte guard
- channel_logo.dart / live_light.dart: num.clamp returns num, which is
  not assignable to int/double contexts (pixels, withAlpha, stops).
- channel_logo_service.dart: subtract replaced entry's bytes on re-store
  so the RAM bound can never inflate.

+++ lib/ui/widgets/channel_logo.dart
@@ -61,7 +61,8 @@ class _ChannelLogoState extends State<ChannelLogo> {
     final int pixels =
         (widget.size * MediaQuery.of(context).devicePixelRatio)
             .round()
-            .clamp(1, 256);
+            .clamp(1, 256)
+            .toInt();
     return SizedBox.square(
       dimension: widget.size,
       child: FutureBuilder<Uint8List?>(
